casting loss

+++ examples/symmetric_conv2d_vae/updated/lightning.py
@@ -65,8 +65,8 @@ class CVAE(pl.LightningModule):
         # recon_x = recon_x.clamp(0, 1)
         x = outputs['expected']
         x = x
-        kld_loss = self.model.kld_loss().half()
-        recon_loss = self.criterion(recon_x.half(), x.half())
+        kld_loss = self.model.kld_loss().float()
+        recon_loss = self.criterion(recon_x.half(), x.half()).float()
         loss = 1.0 * recon_loss + kld_loss
         # loss = recon_loss
         # only use when  on dp
@@ -84,8 +84,8 @@ class CVAE(pl.LightningModule):
         # recon_x = recon_x.clamp(0, 1)
         x = outputs['expected']
         x = x.half()
-        kld_loss = self.model.kld_loss().half()
-        recon_loss = self.criterion(recon_x, x)
+        kld_loss = self.model.kld_loss().float()
+        recon_loss = self.criterion(recon_x, x).float()
         loss = 1.0 * recon_loss + kld_loss
         # loss = recon_loss
         # only use when  on dp
